Convert multiple files at once
I needed to convert a large amount of files into jsons, and i decided it would be quicker to code a simple fix myself.  Now, when you put "all" for the filename, it will just loop through the files at the current directory, and then if it ends in .dat, it will run the code to convert it.

I also added a few error checks that would let it continue to run even if it found an error, which happened a few times, presumably due to corrupted files on my part.